Improvements - shorter

+++ fetchmirrors.sh
@@ -1,9 +1,12 @@
 #!/bin/bash
+### fetchmirrors.sh pacman mirrorlist update utility
+### By: Dylan Schacht (deadhead) deadhead3492@gmail.com
+### Website: http://arch-anywhere.org
+#######################################################
 
 usage() {
 
-	## Display help / usage options menu
-	echo "${Yellow} Usage:${Green} $this <opts> <args> ${Yellow}Ex: ${Green}${this} -s 4 -v -c US"
+	echo "${Yellow} Usage:${Green} $this <opts> <args> ${Yellow}Ex: ${Green}${this} -s 4 -v -c US" # Display help / usage options menu
 	echo
 	echo "${Yellow} Updates pacman mirrorlist directly from archlinux.org"
 	echo
@@ -26,67 +29,48 @@ usage() {
 	echo "${Yellow}Use${Green} $this ${Yellow}without any option to prompt for country code${ColorOff}"
 }
 
-list() {
-	
-	## Display list of country codes to user in terminal
-	echo "${Yellow}Country codes:${ColorOff}"
-	echo "$countries" | column -t
-	echo
-	echo "${Yellow}Note: Use only the upercase two character code in your command ex:${Green} $this -c US"
-	echo "${Yellow}Or simply use:${Green} ${this}${ColorOff}"
-
-}
-
 get_opts() {
-	
-	## Set color variables
+		
+	this=${0##*/} # Set 'this', 'rank_int', 'confirm', 'countries', and color variables
+	rank_int="6"
+	confirm=true
+	countries=( "1) AT Austria - 2) AU  Australia - 3) BE Belgium\n4) BG Bulgaria - 5) BR Brazil - 6) BY Belarus\n7) CA Canada - 8) CL Chile - 9) CN China \n10) CO Columbia - 11) CZ Czech-Republic - 12) DE Germany\n13) DK Denmark - 14) EE Estonia - 15) ES Spain\n16) FI Finland - 17) FR France - 18) GB United-Kingdom\n19) HU Hungary - 20) IE Ireland - 21) IL Isreal\n22) IN India - 23) IT Italy - 24) JP Japan\n25) KR Korea - 26) KZ Kazakhstan - 27) LK Sri-Lanka\n28) LU Luxembourg - 29) LV Lativia - 30) MK Macedonia\n31) NC New-Caledonia - 32) NL Netherlands - 33) NO Norway\n34) NZ New-Zealand - 35) PL Poland - 36) PT Portugal\n37) RO Romania - 38) RS Serbia - 39) RU Russia\n40) SE Sweden - 41) SG Singapore - 42) SK Slovakia\n43) TR Turkey - 44) TW Taiwan - 45) UA Ukraine\n46) US United-States - 47) UZ Uzbekistan - 48) VN Viet-Nam\n49) ZA South-Africa - 50) AM All-Mirrors - 51) AS All-Https" )
 	Green=$'\e[0;32m';
 	Yellow=$'\e[0;33m';
 	Red=$'\e[0;31m';
 	ColorOff=$'\e[0m';
-	
-	## Set 'this', 'rank_int', 'confirm', 'err', and 'countries'
-	this=${0##*/}
-	rank_int="6"
-	confirm=true
-	err=$(echo "${Red}Error: ${Yellow}please select a number from the list.${ColorOff}")
-	countries=$(echo -e "1) AT Austria - 2) AU  Australia - 3) BE Belgium\n4) BG Bulgaria - 5) BR Brazil - 6) BY Belarus\n7) CA Canada - 8) CL Chile - 9) CN China \n10) CO Columbia - 11) CZ Czech-Republic - 12) DE Germany\n13) DK Denmark - 14) EE Estonia - 15) ES Spain\n16) FI Finland - 17) FR France - 18) GB United-Kingdom\n19) HU Hungary - 20) IE Ireland - 21) IL Isreal\n22) IN India - 23) IT Italy - 24) JP Japan\n25) KR Korea - 26) KZ Kazakhstan - 27) LK Sri-Lanka\n28) LU Luxembourg - 29) LV Lativia - 30) MK Macedonia\n31) NC New-Caledonia - 32) NL Netherlands - 33) NO Norway\n34) NZ New-Zealand - 35) PL Poland - 36) PT Portugal\n37) RO Romania - 38) RS Serbia - 39) RU Russia\n40) SE Sweden - 41) SG Singapore - 42) SK Slovakia\n43) TR Turkey - 44) TW Taiwan - 45) UA Ukraine\n46) US United-States - 47) UZ Uzbekistan - 48) VN Viet-Nam\n49) ZA South-Africa - 50) AM All-Mirrors - 51) AS All-Https")
-	
-	while (true)
+
+	while (true) # Loop case statement on 1 parameter until something happens (break || exit 1)
 	  do
 		case "$1" in
-			-d|--nocolor) 
-				Green=$'\e[0m';
-				Yellow=$'\e[0m';
-				Red=$'\e[0m';
-				ColorOff=$'\e[0m';
-				shift
+			-d|--nocolor) # Disable color prompts
+				unset Green Yellow Red ColorOff ; shift
 			;;
-			-q|--noconfirm) 
-				confirm=false
-				shift
+			-q|--noconfirm) # Disable confirmation messages
+				confirm=false ; shift
 			;;
-			-s|--server) 
-				rank_int=$(<<<$*  grep -o "[0-9]\|[0-9][0-9]")
+			-s|--server) # Specify number of servers to add to rank list output
+				rank_int="$2"
 				if ! (<<<"$rank_int" grep "^-\?[0-9]*$" &> /dev/null) || [ -z "$rank_int" ]; then
-					echo "${Red}Error: ${Yellow} invalid number of ranked servers specified ${rank_int}${ColorOff}"
+					echo "${Red}Error: ${Yellow} invalid number of servers specified ${rank_int}${ColorOff}"
 					exit 1
 				fi
 				shift ; shift
 			;;
-			-v|--verbose) 
-				rank_int="$rank_int -v"
-				shift
+			-v|--verbose) # Set verbose switch on
+				rank_int="$rank_int -v" ; shift
 			;;
-			-l|--list) 
-				list
+			-l|--list) # Display a list of country codes to the user
+				echo "${Yellow}Country codes:${ColorOff}"
+				echo -e "$countries" | column -t
+				echo "${Yellow}Note: Use only the upercase two character code in your command ex:${Green} $this -c US"
+				echo "${Yellow}Or simply use:${Green} ${this}${ColorOff}"
 				break
 			;;
-			-h|--help)
-				usage
-				break
+			-h|--help) # Display usage message
+				usage ; break
 			;;
-			-c|--country)
+			-c|--country) # Country parameter allows user to input country code (ex: US)
 				if [ -z "$2" ]; then
 					echo "${Red}Error: ${Yellow}You must enter a country code."
 				elif (<<<"$countries" grep -w "$2" &> /dev/null); then
@@ -99,15 +83,14 @@ get_opts() {
 				fi
 				break
 			;;
-			"")
-				search
-				break
+			"") # Empty 1 parameter means search for country code
+				search ; break
 			;;
-			*)
+			*) # Take anything else as invalid input
 				echo "${Red}Error: ${Yellow}unknown input $1 exiting...${ColorOff}"
 				exit 1
 		esac
-	done
+	done # End case options loop
 	exit
 
 }
@@ -117,12 +100,12 @@ search() {
 	while (true)
 	  do
 		echo "${Yellow}Country codes:${ColorOff}"
-		echo "$countries" | column -t
+		echo -e "$countries" | column -t
 		echo -n "${Yellow}Enter the number corresponding to your country code ${Green}[1,2,3...]${Yellow}:${ColorOff} "
 		read input
 
 		if ! (<<<$input grep "^-\?[0-9]*$" &>/dev/null) || [ "$input" -gt "51" ]; then
-			echo "$err"
+			echo "${Red}Error: ${Yellow}please select a number from the list.${ColorOff}"
 		else
 			break
 		fi
@@ -159,19 +142,18 @@ search() {
 get_list() {
 
 	echo "${Yellow}Fetching new mirrorlist from:${Green} ${query}${ColorOff}"
-	if [ -f /usr/bin/wget ]; then
-		wget -O /tmp/mirrorlist "$query" &> /dev/null
-	else
-		curl -o /tmp/mirrorlist "$query" &> /dev/null
+	
+	if [ -f /usr/bin/wget ]; then wget -O /tmp/mirrorlist "$query" &> /dev/null
+	else curl -o /tmp/mirrorlist "$query" &> /dev/null
 	fi
 	
 	sed -i 's/#//' /tmp/mirrorlist
 	echo "${Yellow}Please wait while ranking${Green} $country_code ${Yellow}mirrors...${ColorOff}"
+	rankmirrors -n "$rank_int" /tmp/mirrorlist > /tmp/mirrorlist.ranked
 	
-	rankmirrors -n $rank_int /tmp/mirrorlist > /tmp/mirrorlist.ranked
 	if [ "$?" -gt "0" ]; then
-		cleanup
 		echo "${Red}Error: ${Yellow}an error occured in ranking mirrorlist exiting..."
+		rm /tmp/{mirrorlist,mirrorlist.ranked} &> /dev/null
 		exit 1
 	fi
 
@@ -196,20 +178,14 @@ get_list() {
 		y|Y|yes|Yes|yY|Yy|yy|YY|"")
 			sudo mv /etc/pacman.d/mirrorlist /etc/pacman.d/mirrorlist.bak
 			sudo mv /tmp/mirrorlist.ranked /etc/pacman.d/mirrorlist
-			cleanup
 			echo "${Green}New mirrorlist installed ${Yellow}- Old mirrorlist backed up to /etc/pacman.d/mirrorlist.bak${ColorOff}"
+			rm /tmp/mirrorlist &> /dev/null
 		;;
 		*)
-			cleanup
 			echo "${Yellow}Mirrorlist was not installed - exiting...${ColorOff}"
+			rm /tmp/{mirrorlist,mirrorlist.ranked} &> /dev/null
 		;;
 	esac
-
-}
-
-cleanup() {
-
-	rm /tmp/{mirrorlist,mirrorlist.ranked} &> /dev/null
 
 }
